+++ package.json
@@ -19,7 +19,7 @@
     "react-scripts": "^4.0.2",
     "react-switch": "^6.0.0",
     "react-tooltip": "^4.2.21",
-    "socket.io-client": "^4.1.2"
+    "socket.io-client": "^4.1.3"
   },
   "devDependencies": {
     "@babel/core": "^7.14.3",
