+++ public/eeg2bidsService.js
@@ -1,8 +1,11 @@
 const {spawn, exec} = require('child_process');
 const os = require('os');
-const pythonLog = require('electron-log');
 const path = require('path');
 const fs = require('fs');
+const {archiveLog} = require('./logs');
+const pythonLog = require('electron-log');
+pythonLog.transports.file.fileName = 'python.log';
+pythonLog.transports.file.archiveLog = archiveLog;
 
 /**
  * EEG2BIDS Wizard Service
@@ -14,20 +17,6 @@ module.exports = class EEG2BIDSService {
   constructor() {
     this.platform = os.platform(); // darwin or win32.
     this.process = null; // the service process.
-    pythonLog.transports.file.fileName = 'python.log';
-    pythonLog.transports.file.archiveLog = (file) => {
-      const today  = new Date();
-      let date = new Date().toISOString().replace(/T/, '-').replaceAll(/:/g, '_').replace(/\..+/, '');
-
-      file = file.toString();
-      const info = path.parse(file);
-
-      try {
-        fs.renameSync(file, path.join(info.dir, info.name + '-' + date + info.ext));
-      } catch (e) {
-        console.warn('Could not rotate log', e);
-      }
-    }
   }
 
   /**
@@ -74,15 +63,17 @@ module.exports = class EEG2BIDSService {
    */
   shutdown(callback) {
     if (this.process) {
-      console.info('[SHUTDOWN of eeg2bidsService]');
+      pythonLog.info('[SHUTDOWN of eeg2bidsService]');
 
       if(os.platform() === 'win32'){
         exec('taskkill /pid ' + this.process.pid + ' /T /F', (error, stdout, stderr) => {
           if (error) {
-            console.error(`exec error: ${error}`);
+            pythonLog.error(`exec error: ${error}`);
           }
-          console.log(stdout);
-          console.error(stderr);
+          pythonLog.log(stdout);
+          if (stderr) {
+            pythonLog.error(stderr);
+          }
           callback();
         });
       } else if(os.platform() !== 'darwin'){
