+++ python/eeg2bids.py
@@ -434,7 +434,7 @@ def eeg_to_bids_thread(data):
             print(traceback.format_exc())
         except WriteError as e:
             response = {
-                'error': 'Cannot write file - ' + str(e)
+                'error': 'Cannot write file - Please delete the ' + data['outputFilename'] + ' folder and try again.'
             }
             print(traceback.format_exc())
         except Exception as e:
